@@ -1,7 +1,7 @@
 {
   "private": true,
   "engines": {
-    "node": "14.x"
+    "node": "16.x"
   },
   "scripts": {
     "build": "tsc -p api/tsconfig.json && tsc -p web/tsconfig.json"
@@ -10,7 +10,8 @@
     "chrome-aws-lambda": "7.0.0",
     "marked": "2.0.0",
     "puppeteer-core": "7.0.0",
-    "twemoji": "13.0.1"
+    "twemoji": "13.0.1",
+    "vercel": "^28.15.3"
   },
   "devDependencies": {
     "@types/marked": "1.2.2",
