@@ -1,16 +1,16 @@
 import { IncomingMessage } from 'http';
 import { parse } from 'url';
-import { ParsedRequest, Theme } from './types';
+import { ParsedRequest } from './types';
 
 export function parseRequest(req: IncomingMessage) {
     console.log('HTTP ' + req.url);
     const { pathname, query } = parse(req.url || '/', true);
-    const { footerURL, images, theme, md, valueHeader, tvl, volumeChange } = (query || {});
+    const { address, theme, md, volume, tvl, type, chainId } = (query || {});
 
     if (Array.isArray(theme)) {
         throw new Error('Expected a single theme');
     }
-    
+
     const arr = (pathname || '/').slice(1).split('.');
     let extension = '';
     let cardName = '';
@@ -23,32 +23,32 @@ export function parseRequest(req: IncomingMessage) {
         cardName = arr.join('.');
     }
 
-    let url = getString(footerURL);
-   
+    cardName = cardName.replace("_", " ");
+
+
     const parsedRequest: ParsedRequest = {
         fileType: extension === 'jpeg' ? extension : 'png',
-        cardName: decodeURIComponent(cardName),
-        valueHeader: getString(valueHeader),
+        address: getString(address),
+        cardName: cardName,
         tvl: getString(tvl),
-        volumeChange: getString(volumeChange),
-        footerURL: decodeURIComponent(url),
+        chainId: getString(chainId),
+        type: getString(type),
+        volume: getString(volume),
         theme: theme === 'dark' ? 'dark' : 'light',
         md: md === '1' || md === 'true',
-        images: getArray(images),
     };
-    parsedRequest.images = getDefaultImages(parsedRequest.images, parsedRequest.theme);
     return parsedRequest;
 }
 
-function getArray(stringOrArray: string[] | string | undefined): string[] {
-    if (typeof stringOrArray === 'undefined') {
-        return [];
-    } else if (Array.isArray(stringOrArray)) {
-        return stringOrArray;
-    } else {
-        return [stringOrArray];
-    }
-}
+// function getArray(stringOrArray: string[] | string | undefined): string[] {
+//     if (typeof stringOrArray === 'undefined') {
+//         return [];
+//     } else if (Array.isArray(stringOrArray)) {
+//         return stringOrArray;
+//     } else {
+//         return [stringOrArray];
+//     }
+// }
 
 function getString(stringOrArray: string[] | string | undefined) {
     if (stringOrArray && Array.isArray(stringOrArray)) {
@@ -62,15 +62,5 @@ function getString(stringOrArray: string[] | string | undefined) {
     }
 }
 
-function getDefaultImages(images: string[], theme: Theme): string[] {
-    const defaultImage = theme === 'light'
-        ? "https://defillama.com/defillama-press-kit/defi/SVG/defillama-dark.svg"
-        : "https://defillama.com/defillama-press-kit/defi/SVG/defillama.svg";
 
-    if (!images || !images[0]) {
-        return [defaultImage];
-    }
-
-    return images;
-}
 
